fix(billing): migrate user_credit_balances + plan_grants to NUMERIC, remove all int() casts
Critical fix: the previous commit migrated ai_operations and
credit_transactions to NUMERIC but MISSED user_credit_balances — the
table where charges actually land. Its used/purchased/monthly_limit
columns were still BIGINT, which would round 0.75 → 1 on charge.

Fixed:
  - Added migrations for user_credit_balances (used, purchased,
    monthly_limit: BIGINT → NUMERIC(20,2))
  - Added migration for plan_credit_grants.monthly_limit (BIGINT → NUMERIC)
  - Updated CREATE TABLE for both (new deployments)
  - Replaced ALL int() casts with float() in billing_service.py:
    get_monthly_remaining, get_total_available, can_afford, _charge_tier,
    reserve_credits, add_purchased_credits, get_user_billing_summary
  - Replaced int() with float() in billing_router.py /balances endpoint

Without this fix, a 0.75 credit charge would silently round to 1,
defeating the purpose of fractional credits.

+++ api/billing_router.py
@@ -132,11 +132,11 @@ async def get_balances(authorization: Optional[str] = Header(None)):
         "balances": [
             {
                 "credit_type": b["credit_type"],
-                "monthly_limit": int(b["monthly_limit"]),
-                "used": int(b["used"]),
+                "monthly_limit": float(b["monthly_limit"]),
+                "used": float(b["used"]),
                 "monthly_remaining": get_monthly_remaining(b),
-                "purchased": int(b.get("purchased", 0)),
-                "total_available": max(0, get_monthly_remaining(b)) + int(b.get("purchased", 0)),
+                "purchased": float(b.get("purchased", 0)),
+                "total_available": max(0, get_monthly_remaining(b)) + float(b.get("purchased", 0)),
                 "reset_date": b.get("reset_date"),
             }
             for b in balances
